Fix issue in mech_driver 'ensure_subports' where binding info is missing
Portions of the mech driver assume that a port's binding information is
a direct attribute of the port; however it's the Neutron ml2 plugin that
actually appends this information to the port. As a result, when
'ensure_subports' queries a port object from Neutron, that port is missing
port binding information needed later on in the function. This change
sets that information for the port.

Change-Id: Ifc4191551df8649816cbbaaaf4d3003b04599f15